% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getICD.R
\name{getICD}
\alias{getICD}
\title{Getting ICD9 codes}
\usage{
getICD(string, ICD)
}
\arguments{
\item{string}{String to match. Can be single term, or vector. Case insensitive.}

\item{ICD}{Which ICD? Specify numeric,  9 or 10.}
}
\value{
If matches are found, returns vector of ICD code.
}
\description{
This function retrieves ICD9/10 codes based on string matching.
}
\keyword{ICD10}
\keyword{ICD9}
