% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getICD9.R
\name{getICD9}
\alias{getICD9}
\title{Getting ICD9 codes}
\usage{
getICD9(refPath, string)
}
\arguments{
\item{refPath}{Path to reference file "all_lkps_maps.xlsx"}

\item{string}{String to match. Can be single term, or vector. Case insensitive.}
}
\description{
This function retrieves ICD9 codes based on string matching.
}
\keyword{ICD9}
