% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getICDDesc.R
\name{getICDDesc}
\alias{getICDDesc}
\title{Getting ICD code descriptions}
\usage{
getICDDesc(code, ICD)
}
\arguments{
\item{code}{Read code to look up. Can be single term, or vector. Returned vector from stringToICD or readToICD function can be used.}

\item{ICD}{Which ICD? Specify numeric,  9 or 10.}
}
\value{
If matches are found, returns data.tablecontaining ICD codes and descriptions.
}
\description{
This function retrieves ICD code descriptionss.
}
\keyword{ICD10}
\keyword{ICD9}
