% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommand.R
\name{getQuery}
\alias{getQuery}
\title{Getting ICD code descriptions}
\usage{
getQuery(table, read2 = NULL, read3 = NULL)
}
\arguments{
\item{table}{"gp_clinical", "gp_scripts" or "gp_registrations"}

\item{read2}{Read codes v2 to look up. Can be single term, or vector. Can be NULL. Returned vector from stringToRead or ICDToRead function can be used.}

\item{read3}{Read codes v3 to look up. Can be single term, or vector. Can be NULL. Returned vector from stringToRead or ICDToRead function can be used.}
}
\value{
SQL query to get data for given read codes.
}
\description{
This function retrieves SQL query to get data for given read codes.
}
\keyword{SQL}
\keyword{query}
