% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReadDescs.R
\name{getReadDesc}
\alias{getReadDesc}
\title{Getting Read code descriptions}
\usage{
getReadDesc(code, read)
}
\arguments{
\item{code}{Read code to look up. Can be single term, or vector. Returned vector from stringToRead or ICDToRead function can be used.}

\item{read}{Which read code version? Specify numeric,  2 or 3.}
}
\value{
If matches are found, returns data.tablecontaining Read codes and descriptions.
}
\description{
This function retrieves read codes descriptions.
}
\keyword{readcodes}
