% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10}
\alias{icd10}
\title{ICD10 codes.}
\format{A data table with two variables:
#' \describe{
  \item{description}{Code Description}
  \item{icd}{ICD10 code}
  ...
}}
\source{
\url{http://biobank.ndph.ox.ac.uk/showcase/showcase/auxdata/primarycare_codings.zip}
}
\usage{
icd10
}
\description{
Mapping ICD10 codes and descriptions
}
\keyword{datasets}
