% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd9}
\alias{icd9}
\title{ICD9 codes.}
\format{A data table with two variables:
#' \describe{
  \item{DESCRIPTION_ICD9}{Code Description}
  \item{ICD9}{ICD9 code}
  ...
}}
\source{
\url{http://biobank.ndph.ox.ac.uk/showcase/showcase/auxdata/primarycare_codings.zip}
}
\usage{
icd9
}
\description{
Mapping ICD9 codes and descriptions
}
\keyword{datasets}
