% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd9Readv3}
\alias{icd9Readv3}
\title{ICD9 codes to read codes v3.}
\format{A data table with two variables:
#' \describe{
  \item{readCode}{Read code v3.}
  \item{icd}{ICD9 code}
  ...
}}
\source{
\url{http://biobank.ndph.ox.ac.uk/showcase/showcase/auxdata/primarycare_codings.zip}
}
\usage{
icd9Readv3
}
\description{
Mapping ICD9 codes to read codes v3.
}
\keyword{datasets}
