% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICDToRead.R
\name{ICDToRead}
\alias{ICDToRead}
\title{Getting Read codes from ICD codes}
\usage{
ICDToRead(code, ICD, read)
}
\arguments{
\item{code}{ICD code to look up. Can be single term, or vector. Returned vector from stringToICD function can be used.}

\item{ICD}{Which ICD? Specify numeric,  9 or 10.}

\item{read}{Which read code version? Specify numeric,  2 or 3.}
}
\value{
If matches are found, returns vector of Read codes.
}
\description{
This function retrieves read codes from ICD codes.
}
\keyword{ICD10}
\keyword{ICD9}
\keyword{readcodes}
