% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getICD9.R
\name{getICD9}
\alias{getICD9}
\title{Getting ICD9 codes}
\usage{
getICD9(string = "")
}
\arguments{
\item{string}{String to match}
}
\description{
This function retrieves ICD9 codes based on string matching.
}
\keyword{ICD9}
